requirements.txt
python-dotenv
flask
requests
yt-dlp
rapidfuzz
youtube-search-python
flask 
flask-socketio
flask-caching
eventlet
questionary
python-slugify
musicbrainzngs
mutagen
pyacoustid
pillow
yt